AwardsShareManagers table

Award voting for managers awards

Usage

data(AwardsShareManagers)

Format

A data frame with 510 observations on the following 7 variables.

awardID

name of award votes were received for

yearID

Year

lgID

League; a factor with levels AL NL

playerID

Manager (player) ID code

pointsWon

Number of points received

pointsMax

Maximum number of points possible

votesFirst

Number of first place votes

Source

Lahman, S. (2025) Lahman's Baseball Database, 1871-2023, 2023 version, https://sabr.org/lahman-database/

Examples

# Voting for the BBWAA Manager of the Year award by year and league

require("dplyr")

# Sort in decreasing order of points by year and league
AwardsShareManagers %>%
   group_by(yearID, lgID) %>%
   arrange(desc(pointsWon))
#> # A tibble: 510 × 7
#> # Groups:   yearID, lgID [82]
#>    awardID                  yearID lgID  playerID pointsWon pointsMax votesFirst
#>    <chr>                     <int> <fct> <chr>        <int>     <int>      <int>
#>  1 BBWAA Manager of the Ye…   2000 NL    bakerdu…       154       160         30
#>  2 BBWAA Manager of the Ye…   2005 NL    coxbo01        152       160         28
#>  3 BBWAA Manager of the Ye…   2011 NL    gibsoki…       152       160         28
#>  4 BBWAA Manager of the Ye…   2009 NL    tracyji…       151       160         29
#>  5 BBWAA Manager of the Ye…   2023 AL    hydebr99       144       150         27
#>  6 BBWAA Manager of the Ye…   2021 NL    kaplega…       143       150         28
#>  7 BBWAA Manager of the Ye…   2004 NL    coxbo01        140       160         22
#>  8 BBWAA Manager of the Ye…   2013 NL    hurdlcl…       140       150         25
#>  9 BBWAA Manager of the Ye…   1991 AL    kellyto…       138       140         27
#> 10 BBWAA Manager of the Ye…   1994 NL    aloufe01       138       140         27
#> # ℹ 500 more rows

# Any unanimous winners?
AwardsShareManagers %>%
  filter(pointsWon == pointsMax)
#> [1] awardID    yearID     lgID       playerID   pointsWon  pointsMax  votesFirst
#> <0 rows> (or 0-length row.names)

# Manager with highest proportion of possible points
AwardsShareManagers %>%
   mutate(propWon = pointsWon/pointsMax) %>%
   arrange(desc(propWon)) %>%
   head(., 1)
#>                     awardID yearID lgID  playerID pointsWon pointsMax
#> 1 BBWAA Manager of the Year   1991   AL kellyto01       138       140
#>   votesFirst   propWon
#> 1         27 0.9857143

# Bobby Cox's MOY vote tallies
AwardsShareManagers %>%
  filter(playerID == "coxbo01")
#>                      awardID yearID lgID playerID pointsWon pointsMax
#> 1  BBWAA Manager of the Year   1983   AL  coxbo01         4        28
#> 2  BBWAA Manager of the Year   1984   AL  coxbo01         9       140
#> 3  BBWAA Manager of the Year   1985   AL  coxbo01       104       140
#> 4  BBWAA Manager of the Year   1991   NL  coxbo01        96       120
#> 5  BBWAA Manager of the Year   1992   NL  coxbo01        29       120
#> 6  BBWAA Manager of the Year   1993   NL  coxbo01        27       140
#> 7  BBWAA Manager of the Year   1994   NL  coxbo01         3       140
#> 8  BBWAA Manager of the Year   1995   NL  coxbo01        20       140
#> 9  BBWAA Manager of the Year   1996   NL  coxbo01        24       140
#> 10 BBWAA Manager of the Year   1997   NL  coxbo01         6       140
#> 11 BBWAA Manager of the Year   1998   NL  coxbo01        17       160
#> 12 BBWAA Manager of the Year   1999   NL  coxbo01        98       160
#> 13 BBWAA Manager of the Year   2000   NL  coxbo01        41       160
#> 14 BBWAA Manager of the Year   2001   NL  coxbo01        12       160
#> 15 BBWAA Manager of the Year   2002   NL  coxbo01        93       160
#> 16 BBWAA Manager of the Year   2003   NL  coxbo01        56       160
#> 17 BBWAA Manager of the Year   2004   NL  coxbo01       140       160
#> 18 BBWAA Manager of the Year   2005   NL  coxbo01       152       160
#> 19 BBWAA Manager of the Year   2009   NL  coxbo01        15       160
#> 20 BBWAA Manager of the Year   2010   NL  coxbo01        28       160
#>    votesFirst
#> 1           4
#> 2           0
#> 3          16
#> 4          13
#> 5           1
#> 6           0
#> 7           0
#> 8           1
#> 9           3
#> 10          0
#> 11          0
#> 12         10
#> 13          1
#> 14          2
#> 15          9
#> 16          6
#> 17         22
#> 18         28
#> 19          0
#> 20          1